fix:move auth schemas to a signle file
