--- host/src/pages/Auth/ForgotPassword/forgotPassword.schema.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-export const ForgotPassWordSchema = z.object({
-    email:z.string().email({message:"Email is required"})
-})
-
-export type ForgotPasswordSchema = z.infer<typeof ForgotPassWordSchema>
