--- host/src/pages/Auth/Login/login.schema.ts
@@ -1,8 +1,0 @@
-import { z } from "zod";
-
-export const LoginSchema = z.object({
-  email: z.string().email({ message: "Email is required" }),
-  password: z.string().min(1, { message: "Password is required" }),
-});
-
-export type LoginSchema = z.infer<typeof LoginSchema>;
